@@ -77,7 +77,7 @@ android {
     }
 
     kotlinOptions {
-        jvmTarget = "11"
+        jvmTarget = "17"
     }
 
     androidResources {
@@ -98,17 +98,17 @@ dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar", "*.aar"], exclude: "proguard-base-7.2.2.jar")
 
     implementation "androidx.appcompat:appcompat:1.6.1"
-    implementation "androidx.recyclerview:recyclerview:1.2.1"
+    implementation "androidx.recyclerview:recyclerview:1.3.0"
     implementation "androidx.cardview:cardview:1.0.0"
     implementation "androidx.viewpager:viewpager:1.0.0"
     implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
-    implementation "com.google.android.material:material:1.8.0"
+    implementation "com.google.android.material:material:1.9.0"
     implementation "com.airbnb.android:lottie:3.1.0"
     implementation "com.google.android.flexbox:flexbox:3.0.0"
 
-    implementation "androidx.annotation:annotation:1.5.0"
-    implementation "com.android.tools:annotations:30.4.1"
-    implementation "com.google.guava:guava:31.1-jre"
+    implementation "androidx.annotation:annotation:1.6.0"
+    implementation "com.android.tools:annotations:31.0.2"
+    implementation "com.google.guava:guava:32.0.1-jre"
 
     def editorGroupId = "io.github.Rosemoe.sora-editor"
     implementation platform("$editorGroupId:bom:0.21.1")
@@ -124,7 +124,7 @@ dependencies {
     minApi26Implementation "com.google.protobuf:protobuf-java:3.19.2"
     //noinspection GradleDependency
     implementation "com.android.tools:sdklib:25.3.0", { exclude group: "com.intellij", module: "annotations" }
-    implementation "com.android.tools:r8:4.0.48"
+    implementation "com.android.tools:r8:8.0.40"
     implementation "com.github.Iyxan23:zipalign-java:1.0.0"
 
     implementation "com.google.code.gson:gson:2.10.1"
@@ -134,12 +134,12 @@ dependencies {
     implementation "com.squareup.okhttp3:okhttp:4.10.0"
     //noinspection GradleDependency
     implementation "com.squareup.okio:okio:3.1.0"
+    implementation 'com.github.Cosmic-Ide:dependencyresolver:593dd3c1b4'
 
-    // TODO: use proguard-core dependency when duplicate class issue has been fixed
     minApi21Implementation "net.sf.proguard:proguard-base:6.0.3"
     minApi26Implementation files('libs/proguard-base-7.2.2.jar')
     minApi26Implementation 'com.guardsquare:proguard-core:9.0.2'
-    minApi26Implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.0'
+    minApi26Implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.22'
 
     //noinspection GradleDependency
     minApi21Implementation "org.eclipse.jdt:ecj:3.13.0"
@@ -157,5 +157,5 @@ dependencies {
     implementation "com.google.firebase:firebase-crashlytics"
     implementation "com.google.firebase:firebase-analytics"
 
-    coreLibraryDesugaring "com.android.tools:desugar_jdk_libs:1.1.5"
+    coreLibraryDesugaring "com.android.tools:desugar_jdk_libs_nio:2.0.3"
 }
