@@ -5,7 +5,6 @@ import static mod.SketchwareUtil.getDip;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.ProgressDialog;
-import android.content.Context;
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnCancelListener;
 import android.content.Intent;
@@ -27,6 +26,8 @@ import android.widget.ImageView;
 import android.widget.PopupMenu;
 import android.widget.Toast;
 
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.appcompat.widget.Toolbar;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
@@ -56,14 +57,15 @@ import com.sketchware.remod.R;
 import com.topjohnwu.superuser.Shell;
 
 import java.io.File;
+import java.lang.ref.WeakReference;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 
 import a.a.a.DB;
-import a.a.a.Dp;
 import a.a.a.GB;
 import a.a.a.MA;
+import a.a.a.ProjectBuilder;
 import a.a.a.ViewEditorFragment;
 import a.a.a.aB;
 import a.a.a.bB;
@@ -111,15 +113,6 @@ import mod.jbk.util.LogUtil;
 import mod.khaled.logcat.LogReaderActivity;
 
 public class DesignActivity extends BaseAppCompatActivity implements OnClickListener {
-
-    private static final int REQUEST_CODE_VIEW_MANAGER = 208;
-    private static final int REQUEST_CODE_IMAGE_MANAGER = 209;
-    private static final int REQUEST_CODE_SOUND_MANAGER = 217;
-    private static final int REQUEST_CODE_LIBRARY_MANAGER = 226;
-    private static final int REQUEST_CODE_FONT_MANAGER = 228;
-    private static final int REQUEST_CODE_COLLECTION_MANAGER = 233;
-    private static final int REQUEST_CODE_SOURCE_CODE_VIEWER = 240;
-
     private ImageView xmlLayoutOrientation;
     private boolean B = false;
     private int currentTabNumber;
@@ -139,6 +132,40 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
     private ViewEditorFragment viewTabAdapter = null;
     private rs eventTabAdapter = null;
     private br componentTabAdapter = null;
+
+    private final ActivityResultLauncher<Intent> openLibraryManager = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
+        if (result.getResultCode() == RESULT_OK) {
+            if (projectFileSelector != null) {
+                projectFileSelector.syncState();
+            }
+            if (viewTabAdapter != null) {
+                viewTabAdapter.n();
+            }
+        }
+    });
+    private final ActivityResultLauncher<Intent> openViewManager = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
+        if (result.getResultCode() == RESULT_OK) {
+            if (viewTabAdapter != null) {
+                viewTabAdapter.i();
+            }
+        }
+    });
+    private final ActivityResultLauncher<Intent> openImageManager = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
+        if (result.getResultCode() == RESULT_OK) {
+            if (projectFileSelector != null) {
+                projectFileSelector.syncState();
+            }
+        }
+    });
+    private final ActivityResultLauncher<Intent> openCollectionManager = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
+        if (result.getResultCode() == RESULT_OK) {
+            viewTabAdapter.j();
+        }
+    });
+    public final ActivityResultLauncher<Intent> changeOpenFile = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), result -> {
+        assert result.getData() != null;
+        projectFileSelector.setXmlFileName(result.getData().getParcelableExtra("project_file"));
+    });
 
     /**
      * Saves the app's version information to the currently opened Sketchware project file.
@@ -179,7 +206,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
      * @param error The error, to be later displayed as text in {@link CompileLogActivity}
      */
     private void indicateCompileErrorOccurred(String error) {
-        new CompileErrorSaver(q.sc_id).writeLogsToFile(error);
+        new CompileErrorSaver(sc_id).writeLogsToFile(error);
         Snackbar snackbar = Snackbar.make(coordinatorLayout, "Show compile log", Snackbar.LENGTH_INDEFINITE);
         snackbar.setAction(Helper.getResString(R.string.common_word_show), v -> {
             if (!mB.a()) {
@@ -274,65 +301,6 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         super.onActivityResult(requestCode, resultCode, data);
 
         switch (requestCode) {
-            case REQUEST_CODE_VIEW_MANAGER:
-                if (resultCode == RESULT_OK) {
-                    if (projectFileSelector != null) {
-                        projectFileSelector.syncState();
-                    }
-                    if (viewTabAdapter != null) {
-                        viewTabAdapter.n();
-                    }
-                }
-                break;
-
-            case REQUEST_CODE_IMAGE_MANAGER:
-                if (resultCode == RESULT_OK) {
-                    if (viewTabAdapter != null) {
-                        viewTabAdapter.i();
-                    }
-                }
-                break;
-
-            case REQUEST_CODE_SOUND_MANAGER:
-            case REQUEST_CODE_FONT_MANAGER:
-                break;
-
-            case 223:
-                if (resultCode == RESULT_OK) {
-                    if (eventTabAdapter != null) {
-                        eventTabAdapter.refreshEvents();
-                    }
-                }
-                break;
-
-            case 224:
-                if (resultCode == RESULT_OK) {
-                    if (componentTabAdapter != null) {
-                        componentTabAdapter.d();
-                    }
-                }
-                break;
-
-            case REQUEST_CODE_LIBRARY_MANAGER:
-                if (resultCode == RESULT_OK) {
-                    if (projectFileSelector != null) {
-                        projectFileSelector.syncState();
-                    }
-                }
-                break;
-
-            case REQUEST_CODE_COLLECTION_MANAGER:
-                if (resultCode == RESULT_OK) {
-                    viewTabAdapter.j();
-                }
-                break;
-
-            case 263:
-                if (resultCode == RESULT_OK) {
-                    projectFileSelector.setXmlFileName(data.getParcelableExtra("project_file"));
-                }
-                break;
-
             case 462:
                 if (resultCode == RESULT_OK && data.getBooleanExtra("req_update_design_activity", false)) {
                     viewTabAdapter.j();
@@ -353,7 +321,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                 viewPager.setCurrentItem(currentTabNumber);
             } else if (t.c("P12I2")) {
                 k();
-                new SaveChangesProjectCloser(getApplicationContext()).execute();
+                new SaveChangesProjectCloser(this).execute();
             } else {
                 showSaveBeforeQuittingDialog();
             }
@@ -364,7 +332,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
     public void onClick(View v) {
         if (!mB.a()) {
             if (v.getId() == R.id.btn_execute) {
-                new BuildAsyncTask(getApplicationContext()).execute();
+                new BuildAsyncTask(this).execute();
             } else if (v.getId() == R.id.btn_compiler_opt) {
                 PopupMenu popupMenu = new PopupMenu(this, findViewById(R.id.btn_compiler_opt));
                 Menu menu = popupMenu.getMenu();
@@ -477,13 +445,13 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                     }
                 }
                 if (componentTabAdapter != null && projectFileBean != null) {
-                    componentTabAdapter.a(projectFileBean);
-                    componentTabAdapter.d();
+                    componentTabAdapter.setProjectFile(projectFileBean);
+                    componentTabAdapter.refreshData();
                 }
             }
         });
         viewPager = findViewById(R.id.viewpager);
-        viewPager.setAdapter(new ViewPagerAdapter(getSupportFragmentManager(), this));
+        viewPager.setAdapter(new ViewPagerAdapter(getSupportFragmentManager()));
         viewPager.setOffscreenPageLimit(3);
         viewPager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
 
@@ -502,7 +470,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                         eventTabAdapter.c();
                     }
                 } else if (currentTabNumber == 2 && componentTabAdapter != null) {
-                    componentTabAdapter.c();
+                    componentTabAdapter.unselectAll();
                 }
                 if (position == 0) {
                     if (viewTabAdapter != null) {
@@ -528,7 +496,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                         projectFileSelector.setFileType(1);
                         projectFileSelector.syncState();
                         if (componentTabAdapter != null) {
-                            componentTabAdapter.d();
+                            componentTabAdapter.refreshData();
                         }
                     }
                 }
@@ -554,7 +522,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
             }
         } else if (itemId == R.id.design_option_menu_title_save_project) {
             k();
-            new ProjectSaver(getApplicationContext()).execute();
+            new ProjectSaver(this).execute();
         }
 
         return super.onOptionsItemSelected(item);
@@ -570,7 +538,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         q = new yq(getApplicationContext(), wq.d(sc_id), projectInfo);
 
         try {
-            new ProjectLoader(getBaseContext(), savedInstanceState).execute();
+            new ProjectLoader(this, savedInstanceState).execute();
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -603,7 +571,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         }
 
         if (!B) {
-            unsavedChangesSaver = new UnsavedChangesSaver(getApplicationContext());
+            unsavedChangesSaver = new UnsavedChangesSaver(this);
             unsavedChangesSaver.execute();
         }
     }
@@ -621,7 +589,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                 dialog.dismiss();
                 try {
                     k();
-                    new SaveChangesProjectCloser(getApplicationContext()).execute();
+                    new SaveChangesProjectCloser(this).execute();
                 } catch (Exception e) {
                     e.printStackTrace();
                     h();
@@ -633,7 +601,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                 dialog.dismiss();
                 try {
                     k();
-                    new DiscardChangesProjectCloser(getApplicationContext()).execute();
+                    new DiscardChangesProjectCloser(this).execute();
                 } catch (Exception e) {
                     e.printStackTrace();
                     h();
@@ -770,7 +738,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
      * Opens {@link ManageCollectionActivity}.
      */
     void toCollectionManager() {
-        launchActivity(ManageCollectionActivity.class, REQUEST_CODE_COLLECTION_MANAGER);
+        launchActivity(ManageCollectionActivity.class, openCollectionManager);
     }
 
     /**
@@ -857,35 +825,35 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
      * Opens {@link ManageFontActivity}.
      */
     void toFontManager() {
-        launchActivity(ManageFontActivity.class, REQUEST_CODE_FONT_MANAGER);
+        launchActivity(ManageFontActivity.class, null);
     }
 
     /**
      * Opens {@link ManageImageActivity}.
      */
     void toImageManager() {
-        launchActivity(ManageImageActivity.class, REQUEST_CODE_IMAGE_MANAGER);
+        launchActivity(ManageImageActivity.class, openImageManager);
     }
 
     /**
      * Opens {@link ManageLibraryActivity}.
      */
     void toLibraryManager() {
-        launchActivity(ManageLibraryActivity.class, REQUEST_CODE_LIBRARY_MANAGER);
+        launchActivity(ManageLibraryActivity.class, openLibraryManager);
     }
 
     /**
      * Opens {@link ManageViewActivity}.
      */
     void toViewManager() {
-        launchActivity(ManageViewActivity.class, REQUEST_CODE_VIEW_MANAGER);
+        launchActivity(ManageViewActivity.class, openViewManager);
     }
 
     /**
      * Opens {@link ManageSoundActivity}.
      */
     void toSoundManager() {
-        launchActivity(ManageSoundActivity.class, REQUEST_CODE_SOUND_MANAGER);
+        launchActivity(ManageSoundActivity.class, null);
     }
 
     /**
@@ -904,11 +872,11 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
             } catch (Exception ignored) {
             }
         }
-        launchActivity(SrcViewerActivity.class, REQUEST_CODE_SOURCE_CODE_VIEWER, new Pair<>("current", current));
+        launchActivity(SrcViewerActivity.class, null, new Pair<>("current", current));
     }
 
     @SafeVarargs
-    private final void launchActivity(Class<? extends Activity> toLaunch, Integer optionalRequestCode, Pair<String, String>... extras) {
+    private final void launchActivity(Class<? extends Activity> toLaunch, ActivityResultLauncher<Intent> optionalLauncher, Pair<String, String>... extras) {
         Intent intent = new Intent(getApplicationContext(), toLaunch);
         intent.setFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP);
         intent.putExtra("sc_id", sc_id);
@@ -916,22 +884,23 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
             intent.putExtra(extra.first, extra.second);
         }
 
-        if (optionalRequestCode == null) {
+        if (optionalLauncher == null) {
             startActivity(intent);
         } else {
-            startActivityForResult(intent, optionalRequestCode);
+            optionalLauncher.launch(intent);
         }
     }
 
-    private class BuildAsyncTask extends MA implements OnCancelListener, BuildProgressReceiver {
-
+    private static class BuildAsyncTask extends MA implements OnCancelListener, BuildProgressReceiver {
+        private final WeakReference<DesignActivity> activity;
         private final BuildingDialog dialog;
         private boolean canceled = false;
 
-        public BuildAsyncTask(Context context) {
-            super(context);
-            addTask(this);
-            dialog = new BuildingDialog(DesignActivity.this);
+        public BuildAsyncTask(DesignActivity activity) {
+            super(activity.getApplicationContext());
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
+            dialog = new BuildingDialog(activity);
             maybeShow();
             dialog.setIsCancelableOnBackPressed(false);
         }
@@ -946,9 +915,10 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         @Override
         public void a() {
             dismiss();
-            runProject.setText(Helper.getResString(R.string.common_word_run));
-            runProject.setClickable(true);
-            getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+            var activity = this.activity.get();
+            activity.runProject.setText(Helper.getResString(R.string.common_word_run));
+            activity.runProject.setClickable(true);
+            activity.getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
         }
 
         /**
@@ -959,12 +929,13 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
          */
         @Override
         public void a(String str) {
-            runOnUiThread(() -> {
+            var activity = this.activity.get();
+            activity.runOnUiThread(() -> {
                 dismiss();
                 SketchwareUtil.toastError("APK build failed");
-                runProject.setText(Helper.getResString(R.string.common_word_run));
-                runProject.setClickable(true);
-                getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+                activity.runProject.setText(Helper.getResString(R.string.common_word_run));
+                activity.runProject.setClickable(true);
+                activity.getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
             });
         }
 
@@ -978,7 +949,10 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
             if (canceled) {
                 cancel(true);
             } else {
+                var activity = this.activity.get();
                 try {
+                    var q = activity.q;
+                    var sc_id = activity.sc_id;
                     publishProgress("Deleting temporary files...");
                     FileUtil.deleteFile(q.projectMyscPath);
 
@@ -998,95 +972,95 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                     kC.c(q.resDirectoryPath + File.separator + "raw");
                     kC = jC.d(sc_id);
                     kC.a(q.assetsPath + File.separator + "fonts");
-                    generateProjectDebugFiles();
+                    activity.generateProjectDebugFiles();
                     q.f();
                     q.e();
 
-                    Dp mDp = new Dp(this, a, q);
+                    ProjectBuilder builder = new ProjectBuilder(this, a, q);
 
-                    mDp.maybeExtractAapt2();
+                    builder.maybeExtractAapt2();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     publishProgress("Extracting built-in libraries...");
-                    mDp.getBuiltInLibrariesReady();
+                    builder.getBuiltInLibrariesReady();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     publishProgress("AAPT2 is running...");
-                    mDp.compileResources();
+                    builder.compileResources();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
-                    KotlinCompilerBridge.compileKotlinCodeIfPossible(this, mDp);
+                    KotlinCompilerBridge.compileKotlinCodeIfPossible(this, builder);
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     publishProgress("Java is compiling...");
-                    mDp.compileJavaCode();
+                    builder.compileJavaCode();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     /* Encrypt Strings in classes if enabled */
-                    StringfogHandler stringfogHandler = new StringfogHandler(q.sc_id);
-                    stringfogHandler.start(this, mDp);
+                    StringfogHandler stringfogHandler = new StringfogHandler(sc_id);
+                    stringfogHandler.start(this, builder);
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     /* Obfuscate classes if enabled */
-                    ProguardHandler proguardHandler = new ProguardHandler(q.sc_id);
-                    proguardHandler.start(this, mDp);
+                    ProguardHandler proguardHandler = new ProguardHandler(sc_id);
+                    proguardHandler.start(this, builder);
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
-                    publishProgress(mDp.getDxRunningText());
-                    mDp.createDexFilesFromClasses();
+                    publishProgress(builder.getDxRunningText());
+                    builder.createDexFilesFromClasses();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     publishProgress("Merging DEX files...");
-                    mDp.getDexFilesReady();
+                    builder.getDexFilesReady();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     publishProgress("Building APK...");
-                    mDp.buildApk();
+                    builder.buildApk();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
                     publishProgress("Signing APK...");
-                    mDp.signDebugApk();
+                    builder.signDebugApk();
                     if (canceled) {
                         cancel(true);
                         return;
                     }
 
-                    installBuiltApk();
+                    activity.installBuiltApk();
                 } catch (MissingFileException e) {
-                    runOnUiThread(() -> {
+                    activity.runOnUiThread(() -> {
                         boolean isMissingDirectory = e.isMissingDirectory();
 
-                        aB dialog = new aB(DesignActivity.this);
+                        aB dialog = new aB(activity);
                         if (isMissingDirectory) {
                             dialog.b("Missing directory detected");
                             dialog.a("A directory important for building is missing. " +
@@ -1109,7 +1083,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
                     });
                 } catch (Throwable tr) {
                     LogUtil.e("DesignActivity$BuildAsyncTask", "Failed to build project", tr);
-                    indicateCompileErrorOccurred(tr instanceof zy ? tr.getMessage() : Log.getStackTraceString(tr));
+                    activity.indicateCompileErrorOccurred(tr instanceof zy ? tr.getMessage() : Log.getStackTraceString(tr));
                 }
             }
         }
@@ -1118,7 +1092,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
          * Dismiss this dialog, if DesignActivity hasn't been destroyed.
          */
         private void dismiss() {
-            if (!isDestroyed()) {
+            if (!activity.get().isDestroyed()) {
                 if (dialog.isShowing()) {
                     dialog.dismiss();
                 }
@@ -1131,7 +1105,7 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
          * @param progressText The new text to display as progress
          */
         public void setProgress(String progressText) {
-            runOnUiThread(() -> {
+            activity.get().runOnUiThread(() -> {
                 if (dialog.isShowing()) {
                     dialog.setProgress(progressText);
                 }
@@ -1161,10 +1135,11 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         @Override
         public void onCancelled() {
             super.onCancelled();
-            runOnUiThread(() -> {
-                runProject.setText(Helper.getResString(R.string.common_word_run));
-                runProject.setClickable(true);
-                getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+            var activity = this.activity.get();
+            activity.runOnUiThread(() -> {
+                activity.runProject.setText(Helper.getResString(R.string.common_word_run));
+                activity.runProject.setClickable(true);
+                activity.getWindow().clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
             });
             dismiss();
         }
@@ -1177,10 +1152,11 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         @Override
         public void onPreExecute() {
             super.onPreExecute();
-            runProject.setText("Building APK file...");
-            runProject.setClickable(false);
-            r.a("P1I10", true);
-            getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+            var activity = this.activity.get();
+            activity.runProject.setText("Building APK file...");
+            activity.runProject.setClickable(false);
+            activity.r.a("P1I10", true);
+            activity.getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
         }
 
         @Override
@@ -1189,42 +1165,44 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         }
     }
 
-    private class ProjectLoader extends MA {
-
+    private static class ProjectLoader extends MA {
+        private final WeakReference<DesignActivity> activity;
         private final Bundle savedInstanceState;
 
-        public ProjectLoader(Context context, Bundle savedInstanceState) {
-            super(context);
-            addTask(this);
+        public ProjectLoader(DesignActivity activity, Bundle savedInstanceState) {
+            super(activity.getBaseContext());
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
             this.savedInstanceState = savedInstanceState;
         }
 
         @Override
         public void a() {
+            var activity = this.activity.get();
             if (savedInstanceState != null) {
-                projectFileSelector.onRestoreInstanceState(savedInstanceState);
+                activity.projectFileSelector.onRestoreInstanceState(savedInstanceState);
                 if (savedInstanceState.getInt("file_selector_current_file_type") == 0) {
-                    xmlLayoutOrientation.setVisibility(View.VISIBLE);
+                    activity.xmlLayoutOrientation.setVisibility(View.VISIBLE);
                 } else {
-                    xmlLayoutOrientation.setVisibility(View.GONE);
+                    activity.xmlLayoutOrientation.setVisibility(View.GONE);
                 }
             }
 
-            projectFileSelector.syncState();
-            h();
+            activity.projectFileSelector.syncState();
+            activity.h();
             if (savedInstanceState == null) {
-                checkForUnsavedProjectData();
+                activity.checkForUnsavedProjectData();
             }
         }
 
         @Override
         public void a(String str) {
-            h();
+            activity.get().h();
         }
 
         @Override
         public void b() {
-            loadProject(savedInstanceState != null);
+            activity.get().loadProject(savedInstanceState != null);
         }
 
         @Override
@@ -1233,27 +1211,32 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         }
     }
 
-    private class DiscardChangesProjectCloser extends MA {
+    private static class DiscardChangesProjectCloser extends MA {
+        private final WeakReference<DesignActivity> activity;
 
-        public DiscardChangesProjectCloser(Context context) {
-            super(context);
-            addTask(this);
+        public DiscardChangesProjectCloser(DesignActivity activity) {
+            super(activity.getApplicationContext());
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
         }
 
         @Override
         public void a() {
-            h();
-            finish();
+            var activity = this.activity.get();
+            activity.h();
+            activity.finish();
         }
 
         @Override
         public void a(String str) {
-            h();
-            finish();
+            var activity = this.activity.get();
+            activity.h();
+            activity.finish();
         }
 
         @Override
         public void b() {
+            var sc_id = activity.get().sc_id;
             jC.d(sc_id).v();
             jC.d(sc_id).w();
             jC.d(sc_id).u();
@@ -1265,18 +1248,22 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         }
     }
 
-    private class ProjectSaver extends MA {
+    private static class ProjectSaver extends MA {
+        private final WeakReference<DesignActivity> activity;
 
-        public ProjectSaver(Context context) {
-            super(context);
-            addTask(this);
+        public ProjectSaver(DesignActivity activity) {
+            super(activity.getApplicationContext());
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
         }
 
         @Override
         public void a() {
             bB.a(a, Helper.getResString(R.string.common_message_complete_save), bB.TOAST_NORMAL).show();
-            saveVersionCodeInformationToProject();
-            h();
+            var activity = this.activity.get();
+            var sc_id = activity.sc_id;
+            activity.saveVersionCodeInformationToProject();
+            activity.h();
             jC.d(sc_id).f();
             jC.d(sc_id).g();
             jC.d(sc_id).e();
@@ -1285,11 +1272,12 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         @Override
         public void a(String str) {
             bB.b(a, Helper.getResString(R.string.common_error_failed_to_save), bB.TOAST_NORMAL).show();
-            h();
+            activity.get().h();
         }
 
         @Override
         public void b() {
+            var sc_id = activity.get().sc_id;
             jC.d(sc_id).a();
             jC.b(sc_id).m();
             jC.a(sc_id).j();
@@ -1303,29 +1291,33 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         }
     }
 
-    private class SaveChangesProjectCloser extends MA {
+    private static class SaveChangesProjectCloser extends MA {
+        private final WeakReference<DesignActivity> activity;
 
-        public SaveChangesProjectCloser(Context context) {
-            super(context);
-            addTask(this);
+        public SaveChangesProjectCloser(DesignActivity activity) {
+            super(activity.getApplicationContext());
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
         }
 
         @Override
         public void a() {
             bB.a(a, Helper.getResString(R.string.common_message_complete_save), bB.TOAST_NORMAL).show();
-            saveVersionCodeInformationToProject();
-            h();
-            finish();
+            var activity = this.activity.get();
+            activity.saveVersionCodeInformationToProject();
+            activity.h();
+            activity.finish();
         }
 
         @Override
         public void a(String str) {
             bB.b(a, Helper.getResString(R.string.common_error_failed_to_save), bB.TOAST_NORMAL).show();
-            h();
+            activity.get().h();
         }
 
         @Override
         public void b() {
+            var sc_id = activity.get().sc_id;
             jC.d(sc_id).a();
             jC.b(sc_id).m();
             jC.a(sc_id).j();
@@ -1340,11 +1332,13 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
         }
     }
 
-    private class UnsavedChangesSaver extends MA {
+    private static class UnsavedChangesSaver extends MA {
+        private final WeakReference<DesignActivity> activity;
 
-        public UnsavedChangesSaver(Context context) {
-            super(context);
-            addTask(this);
+        public UnsavedChangesSaver(DesignActivity activity) {
+            super(activity.getApplicationContext());
+            this.activity = new WeakReference<>(activity);
+            activity.addTask(this);
         }
 
         @Override
@@ -1357,26 +1351,20 @@ public class DesignActivity extends BaseAppCompatActivity implements OnClickList
 
         @Override
         public void b() {
-            jC.a(sc_id).k();
+            jC.a(activity.get().sc_id).k();
         }
 
         @Override
         protected String doInBackground(Void... voids) {
             return a(voids);
         }
-
-        @Override
-        public void onCancelled() {
-            super.onCancelled();
-        }
     }
 
     private class ViewPagerAdapter extends FragmentPagerAdapter {
-
         private final String[] labels;
 
-        public ViewPagerAdapter(FragmentManager xf, Context context) {
-            super(xf);
+        public ViewPagerAdapter(FragmentManager fragmentManager) {
+            super(fragmentManager);
             labels = new String[]{
                     Helper.getResString(R.string.design_tab_title_view),
                     Helper.getResString(R.string.design_tab_title_event),
