@@ -3,11 +3,10 @@ package mod.hey.studios.project.proguard;
 import com.google.gson.Gson;
 
 import java.io.IOException;
-import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.HashMap;
 
-import a.a.a.Dp;
+import a.a.a.ProjectBuilder;
 import mod.agus.jcoderz.lib.FileUtil;
 import mod.hey.studios.util.Helper;
 import mod.jbk.build.BuildProgressReceiver;
@@ -207,10 +206,10 @@ public class ProguardHandler {
         FileUtil.writeFile(fm_config_path, new Gson().toJson(fullModeLibs));
     }
 
-    public void start(BuildProgressReceiver progressReceiver, Dp dp) throws IOException {
+    public void start(BuildProgressReceiver progressReceiver, ProjectBuilder builder) throws IOException {
         if (isProguardEnabled()) {
             progressReceiver.onProgress("ProGuarding classes...");
-            dp.runProguard();
+            builder.runProguard();
         }
     }
 }
