@@ -1,16 +1,15 @@
 package mod.hey.studios.compiler.kotlin;
 
 import a.a.a.BuiltInLibraryManager;
-import a.a.a.Dp;
+import a.a.a.ProjectBuilder;
 import a.a.a.yq;
 import mod.jbk.build.BuildProgressReceiver;
 
 public class KotlinCompilerBridge {
-
-    public static void compileKotlinCodeIfPossible(BuildProgressReceiver task, Dp dp) throws Throwable {
+    public static void compileKotlinCodeIfPossible(BuildProgressReceiver task, ProjectBuilder builder) throws Throwable {
     }
 
-    public static void maybeAddKotlinBuiltInLibraryDependenciesIfPossible(Dp dp, BuiltInLibraryManager builtInLibraryManager) {
+    public static void maybeAddKotlinBuiltInLibraryDependenciesIfPossible(ProjectBuilder builder, BuiltInLibraryManager builtInLibraryManager) {
     }
 
     public static void maybeAddKotlinFilesToClasspath(StringBuilder classpath, yq workspace) {
